chore: moved 'spanned' from parser to common
