+++ crates/common/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod pool;
+mod span;
+pub use span::*;
 pub mod symbols;
 pub use symbols::*;
 
@@ -16,20 +18,6 @@ pub enum VisibilityModifier {
     ParentPublic,
 }
 
-///The representation of the bounds of something on the code.
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Hash)]
-pub struct Span {
-    pub start: u32,
-    pub end: u32,
-}
-
-impl Span {
-    ///Merges this span with the given `target`. The returned span will have the initial position of this one, and the final position of the given `target`
-    pub fn merge_with(mut self, target: Self) -> Self {
-        self.end = target.end;
-        self
-    }
-}
 ///Some operator on the code. Something like, +, - , *, /, &, &&, etc
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
 pub enum Operator {
